niet meer voor image versie > v2.3

+++ install/git2as.sh
@@ -18,12 +18,15 @@ then
   cp -r /opt/SCAPE604/git/apri-sensor/apri-sensor-connector /opt/SCAPE604/apri-sensor/.
   cp -r /opt/SCAPE604/git/apri-sensor/apri-sensor-raspi /opt/SCAPE604/apri-sensor/.
   cp -r /opt/SCAPE604/git/apri-sensor/apri-sensor-redis /opt/SCAPE604/apri-sensor/.
-  cp -r /opt/SCAPE604/git/apri-sensor/aprisensor-netmanager-runtime /opt/SCAPE604/apri-sensor/.
   cp -r /opt/SCAPE604/git/apri-sensor/aprisensor-nmcli /opt/SCAPE604/apri-sensor/.
   cp -r /opt/SCAPE604/git/apri-sensor/images /opt/SCAPE604/apri-sensor/.
   cp -r /opt/SCAPE604/git/apri-sensor/apri-config/aprisensor-types /opt/SCAPE604/apri-sensor/apri-config/aprisensor-types
-# sudo cp -r /opt/SCAPE604/git/apri-sensor/apri-sensor-bam1020 /opt/SCAPE604/apri-sensor/.
-# sudo cp /opt/SCAPE604/git/apri-sensor/apri-config/apri-config.js /opt/SCAPE604/apri-sensor/apri-config/.
   sudo systemctl restart SCAPE604-aprisensor-nmcli
-  echo "git2as files copied"
-fi
+  # sudo cp -r /opt/SCAPE604/git/apri-sensor/apri-sensor-bam1020 /opt/SCAPE604/apri-sensor/.
+  # sudo cp /opt/SCAPE604/git/apri-sensor/apri-config/apri-config.js /opt/SCAPE604/apri-sensor/apri-config/.
+  ## niet meer voor image versie > v2.3
+  if [ -f /opt/SCAPE604/git/apri-sensor/aprisensor-netmanager-runtime ]
+  then
+    cp -r /opt/SCAPE604/git/apri-sensor/aprisensor-netmanager-runtime /opt/SCAPE604/apri-sensor/.
+  fi
+echo "git2as files copied"
